+++ src/python/zensols/config/__init__.py
@@ -33,6 +33,6 @@ from .condyaml import *
 from .importini import *
 from .envconfig import *
 from .treeimpmod import *
-from .treeconfig import *
+from .importtree import *
 from .keychain import *
 from .meta import *